requirements.txt
Django==4.0.6
folium==0.14.0
gunicorn==20.1.0
pandas==2.1.4
whitenoise==6.2.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
google-generativeai==0.3.2
PyMuPdf==1.23.8
pytesseract==0.3.10
Pillow==10.1.0
python-dotenv==1.0.0
opencv-python==4.9.0.80
assemblyai==0.20.2
langchain==0.1.0
langchain_experimental==0.0.49
openai==1.7.1
tabulate==0.9.0
InstructorEmbedding
torch
sentence-transformers
chromadb